requirements.txt
flask~=3.1.0
requests~=2.32.3
pyyaml~=6.0.2
langchain-openai~=0.3.8
numpy~=2.0.2
tqdm~=4.67.1
FlagEmbedding~=1.3.4
nltk~=3.9.1
ragas~=0.2.14

datasets~=3.6.0
openai~=1.78.0
chromadb~=1.0.8
streamlit~=1.45.0
langchain~=0.3.25
torch~=2.7.0 #+cu126
transformers~=4.51.3
matplotlib~=3.9.4
cohere~=5.15.0